Refactor to a single sample using many backends
To avoid code duplication, the Azure-specific sample notebook was
removed and the existing sample was modified to contain changes
required to use different backends.
